fix merkleAccumulator
