+++ arbos/merkleAccumulator/merkleAccumulator.go
@@ -6,6 +6,7 @@ package merkleAccumulator
 
 import (
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/offchainlabs/arbstate/arbos/storage"
 	util_math "github.com/offchainlabs/arbstate/util"
 )
@@ -62,6 +63,22 @@ func (acc *MerkleAccumulator) NonPersistentClone() (*MerkleAccumulator, error) {
 	}
 	mbu := &storage.MemoryBackedUint64{}
 	return &MerkleAccumulator{nil, mbu, partials}, mbu.Set(size)
+}
+
+func (acc *MerkleAccumulator) Keccak(data ...[]byte) ([]byte, error) {
+	if acc.backingStorage != nil {
+		return acc.backingStorage.Keccak(data...)
+	} else {
+		return crypto.Keccak256(data...), nil
+	}
+}
+
+func (acc *MerkleAccumulator) KeccakHash(data ...[]byte) (common.Hash, error) {
+	if acc.backingStorage != nil {
+		return acc.backingStorage.KeccakHash(data...)
+	} else {
+		return crypto.Keccak256Hash(data...), nil
+	}
 }
 
 func (acc *MerkleAccumulator) getPartial(level uint64) (*common.Hash, error) {
@@ -131,7 +148,7 @@ func (acc *MerkleAccumulator) Append(itemHash common.Hash) ([]MerkleTreeNodeEven
 			err := acc.setPartial(level, &h)
 			return events, err
 		}
-		soFar, err = acc.backingStorage.Keccak(thisLevel.Bytes(), soFar)
+		soFar, err = acc.Keccak(thisLevel.Bytes(), soFar)
 		if err != nil {
 			return nil, err
 		}
@@ -169,14 +186,14 @@ func (acc *MerkleAccumulator) Root() (common.Hash, error) {
 				capacityInHash = capacity
 			} else {
 				for capacityInHash < capacity {
-					h, err := acc.backingStorage.KeccakHash(hashSoFar.Bytes(), make([]byte, 32))
+					h, err := acc.KeccakHash(hashSoFar.Bytes(), make([]byte, 32))
 					if err != nil {
 						return common.Hash{}, err
 					}
 					hashSoFar = &h
 					capacityInHash *= 2
 				}
-				h, err := acc.backingStorage.KeccakHash(partial.Bytes(), hashSoFar.Bytes())
+				h, err := acc.KeccakHash(partial.Bytes(), hashSoFar.Bytes())
 				if err != nil {
 					return common.Hash{}, err
 				}
