+++ util/merkletree/merkleEventProof_test.go
@@ -47,3 +47,34 @@ func TestProofForNext(t *testing.T) {
 		}
 	}
 }
+
+func ProofFromAccumulator(acc *merkleAccumulator.MerkleAccumulator, nextHash common.Hash) (*MerkleProof, error) {
+	origPartials, err := acc.GetPartials()
+	if err != nil {
+		return nil, err
+	}
+	partials := make([]common.Hash, len(origPartials))
+	for i, orig := range origPartials {
+		partials[i] = *orig
+	}
+	clone, err := acc.NonPersistentClone()
+	if err != nil {
+		return nil, err
+	}
+	_, err = clone.Append(nextHash)
+	if err != nil {
+		return nil, err
+	}
+	root, _ := clone.Root()
+	size, err := acc.Size()
+	if err != nil {
+		return nil, err
+	}
+
+	return &MerkleProof{
+		RootHash:  root,
+		LeafHash:  nextHash,
+		LeafIndex: size,
+		Proof:     partials,
+	}, nil
+}
